// src/types.ts
export type MaybeReactive<T> = T | (() => T);

// Extract event handlers (onXXX keys) from an attributes interface
type EventHandlerKeys<T> = {
  [K in keyof T]: K extends `on${string}` ? K : never;
}[keyof T];

// Non-event attribute keys
export type AttrKeys<T> = Exclude<keyof T, EventHandlerKeys<T>>;

// Make attribute values optionally reactive (exclude event handlers)
export type ReactiveAttrs<T> = {
  [K in AttrKeys<T>]?: T[K] extends infer V ? MaybeReactive<V> : never;
} & {
  [K in EventHandlerKeys<T>]?: T[K];
} & {
  [key: string]: unknown;
};

// Prop proxy: each property accepts plain value or reactive
type ReactiveProp<E> = {
  [K in keyof E]: MaybeReactive<E[K]>;
};

// Tools passed to the builder callback, strongly typed per element
export interface TagTools<E extends HTMLElement> {
  el: E;
  on: <K extends keyof HTMLElementEventMap>(
    event: K,
    handler: (this: E, ev: HTMLElementEventMap[K]) => any,
    options?: boolean | AddEventListenerOptions,
  ) => void;
  off: <K extends keyof HTMLElementEventMap>(
    event: K,
    handler: (this: E, ev: HTMLElementEventMap[K]) => any,
    options?: boolean | EventListenerOptions,
  ) => void;
  prop: ReactiveProp<E>;
  attr: {
    [K in AttrKeys<JSX.HTMLAttributes>]?: JSX.HTMLAttributes[K];
  } & Record<string, any>;
}

// Map from IntrinsicElements key to the corresponding HTMLElement type
type ElementTypeFor<K extends keyof IntrinsicElements> =
  K extends keyof HTMLElementTagNameMap ? HTMLElementTagNameMap[K] : HTMLElement;

// The tag proxy type: for each known element, provides strongly-typed overloads
export type TagProxy = {
  [K in keyof IntrinsicElements]: ((
    attrs: ReactiveAttrs<IntrinsicElements[K]>,
    builder?: (tools: TagTools<ElementTypeFor<K>>) => void,
  ) => ElementTypeFor<K>) &
    ((
      builder?: (tools: TagTools<ElementTypeFor<K>>) => void,
    ) => ElementTypeFor<K>);
} & {
  [K: string]: ((
    attrs: ReactiveAttrs<JSX.HTMLAttributes>,
    builder?: (tools: TagTools<HTMLElement>) => void,
  ) => HTMLElement) &
    ((
      builder?: (tools: TagTools<HTMLElement>) => void,
    ) => HTMLElement);
};

export declare namespace JSX {
// ...
    export interface HTMLAttributes extends JSXAttributes, EventAttributes {
        accesskey?: string | undefined;
        autocapitalize?: 'off' | 'none' | 'on' | 'sentences' | 'words' | 'characters' | undefined;
        autofocus?: boolean | undefined;
        class?: string | Promise<string> | undefined;
        contenteditable?: boolean | 'inherit' | 'plaintext-only' | undefined;
        contextmenu?: string | undefined;
        dir?: string | undefined;
        draggable?: 'true' | 'false' | boolean | undefined;
        enterkeyhint?: 'enter' | 'done' | 'go' | 'next' | 'previous' | 'search' | 'send' | undefined;
        hidden?: boolean | undefined;
        id?: string | undefined;
        inert?: boolean | undefined;
        inputmode?: 'none' | 'text' | 'tel' | 'url' | 'email' | 'numeric' | 'decimal' | 'search' | undefined;
        is?: string | undefined;
        itemid?: string | undefined;
        itemprop?: string | undefined;
        itemref?: string | undefined;
        itemscope?: boolean | undefined;
        itemtype?: string | undefined;
        lang?: string | undefined;
        nonce?: string | undefined;
        placeholder?: string | undefined;
        /** @see https://developer.mozilla.org/en-US/docs/Web/API/Popover_API */
        popover?: boolean | 'auto' | 'manual' | undefined;
        slot?: string | undefined;
        spellcheck?: boolean | undefined;
        style?: CSSProperties | string | undefined;
        tabindex?: number | undefined;
        title?: string | undefined;
        translate?: 'yes' | 'no' | undefined;
        itemProp?: string | undefined;
    }
    type HTMLAttributeReferrerPolicy = '' | 'no-referrer' | 'no-referrer-when-downgrade' | 'origin' | 'origin-when-cross-origin' | 'same-origin' | 'strict-origin' | 'strict-origin-when-cross-origin' | 'unsafe-url';
    type HTMLAttributeAnchorTarget = StringLiteralUnion<'_self' | '_blank' | '_parent' | '_top'>;
    interface AnchorHTMLAttributes extends HTMLAttributes {
        download?: string | boolean | undefined;
        href?: string | undefined;
        hreflang?: string | undefined;
        media?: string | undefined;
        ping?: string | undefined;
        target?: HTMLAttributeAnchorTarget | undefined;
        type?: StringLiteralUnion<BaseMime> | undefined;
        referrerpolicy?: HTMLAttributeReferrerPolicy | undefined;
    }
    interface AudioHTMLAttributes extends MediaHTMLAttributes {
    }
    interface AreaHTMLAttributes extends HTMLAttributes {
        alt?: string | undefined;
        coords?: string | undefined;
        download?: string | boolean | undefined;
        href?: string | undefined;
        hreflang?: string | undefined;
        media?: string | undefined;
        referrerpolicy?: HTMLAttributeReferrerPolicy | undefined;
        shape?: string | undefined;
        target?: HTMLAttributeAnchorTarget | undefined;
    }
    interface BaseHTMLAttributes extends HTMLAttributes {
        href?: string | undefined;
        target?: HTMLAttributeAnchorTarget | undefined;
    }
    interface BlockquoteHTMLAttributes extends HTMLAttributes {
        cite?: string | undefined;
    }
    /** @see https://developer.mozilla.org/en-US/docs/Web/API/Popover_API */
    type HTMLAttributePopoverTargetAction = 'show' | 'hide' | 'toggle';
    interface ButtonHTMLAttributes extends HTMLAttributes {
        disabled?: boolean | undefined;
        form?: string | undefined;
        formenctype?: HTMLAttributeFormEnctype | undefined;
        formmethod?: HTMLAttributeFormMethod | undefined;
        formnovalidate?: boolean | undefined;
        formtarget?: HTMLAttributeAnchorTarget | undefined;
        name?: string | undefined;
        type?: 'submit' | 'reset' | 'button' | undefined;
        value?: string | ReadonlyArray<string> | number | undefined;
        popovertarget?: string | undefined;
        popovertargetaction?: HTMLAttributePopoverTargetAction | undefined;
        formAction?: string | Function | undefined;
    }
    interface CanvasHTMLAttributes extends HTMLAttributes {
        height?: number | string | undefined;
        width?: number | string | undefined;
    }
    interface ColHTMLAttributes extends HTMLAttributes {
        span?: number | undefined;
        width?: number | string | undefined;
    }
    interface ColgroupHTMLAttributes extends HTMLAttributes {
        span?: number | undefined;
    }
    interface DataHTMLAttributes extends HTMLAttributes {
        value?: string | ReadonlyArray<string> | number | undefined;
    }
    interface DetailsHTMLAttributes extends HTMLAttributes {
        open?: boolean | undefined;
    }
    interface DelHTMLAttributes extends HTMLAttributes {
        cite?: string | undefined;
        dateTime?: string | undefined;
    }
    interface DialogHTMLAttributes extends HTMLAttributes {
        open?: boolean | undefined;
    }
    interface EmbedHTMLAttributes extends HTMLAttributes {
        height?: number | string | undefined;
        src?: string | undefined;
        type?: StringLiteralUnion<BaseMime> | undefined;
        width?: number | string | undefined;
    }
    interface FieldsetHTMLAttributes extends HTMLAttributes {
        disabled?: boolean | undefined;
        form?: string | undefined;
        name?: string | undefined;
    }
// ...
    interface FormHTMLAttributes extends HTMLAttributes {
        'accept-charset'?: StringLiteralUnion<'utf-8'> | undefined;
        autocomplete?: HTMLAttributeFormAutocomplete | undefined;
        enctype?: HTMLAttributeFormEnctype | undefined;
        method?: HTMLAttributeFormMethod | undefined;
        name?: string | undefined;
        novalidate?: boolean | undefined;
        target?: HTMLAttributeAnchorTarget | undefined;
        action?: string | Function | undefined;
    }
    interface HtmlHTMLAttributes extends HTMLAttributes {
        manifest?: string | undefined;
    }
    interface IframeHTMLAttributes extends HTMLAttributes {
        allow?: string | undefined;
        allowfullscreen?: boolean | undefined;
        height?: number | string | undefined;
        loading?: 'eager' | 'lazy' | undefined;
        name?: string | undefined;
        referrerpolicy?: HTMLAttributeReferrerPolicy | undefined;
        sandbox?: string | undefined;
        seamless?: boolean | undefined;
        src?: string | undefined;
        srcdoc?: string | undefined;
        width?: number | string | undefined;
    }
// ...
    interface InsHTMLAttributes extends HTMLAttributes {
        cite?: string | undefined;
        datetime?: string | undefined;
    }
// ...
    interface KeygenHTMLAttributes extends HTMLAttributes {
        challenge?: string | undefined;
        disabled?: boolean | undefined;
        form?: string | undefined;
        keytype?: string | undefined;
        name?: string | undefined;
    }
    interface LabelHTMLAttributes extends HTMLAttributes {
        form?: string | undefined;
        for?: string | undefined;
    }
// ...
    export interface IntrinsicElements {
        a: AnchorHTMLAttributes;
        abbr: HTMLAttributes;
        address: HTMLAttributes;
        area: AreaHTMLAttributes;
        article: HTMLAttributes;
        aside: HTMLAttributes;
        audio: AudioHTMLAttributes;
        b: HTMLAttributes;
        base: BaseHTMLAttributes;
        bdi: HTMLAttributes;
        bdo: HTMLAttributes;
        big: HTMLAttributes;
        blockquote: BlockquoteHTMLAttributes;
        body: HTMLAttributes;
        br: HTMLAttributes;
        button: ButtonHTMLAttributes;
        canvas: CanvasHTMLAttributes;
        caption: HTMLAttributes;
        center: HTMLAttributes;
        cite: HTMLAttributes;
        code: HTMLAttributes;
        col: ColHTMLAttributes;
        colgroup: ColgroupHTMLAttributes;
        data: DataHTMLAttributes;
        datalist: HTMLAttributes;
        dd: HTMLAttributes;
        del: DelHTMLAttributes;
        details: DetailsHTMLAttributes;
        dfn: HTMLAttributes;
        dialog: DialogHTMLAttributes;
        div: HTMLAttributes;
        dl: HTMLAttributes;
        dt: HTMLAttributes;
        em: HTMLAttributes;
        embed: EmbedHTMLAttributes;
        fieldset: FieldsetHTMLAttributes;
        figcaption: HTMLAttributes;
        figure: HTMLAttributes;
        footer: HTMLAttributes;
        form: FormHTMLAttributes;
        h1: HTMLAttributes;
        h2: HTMLAttributes;
        h3: HTMLAttributes;
        h4: HTMLAttributes;
        h5: HTMLAttributes;
        h6: HTMLAttributes;
        head: HTMLAttributes;
        header: HTMLAttributes;
        hgroup: HTMLAttributes;
        hr: HTMLAttributes;
        html: HtmlHTMLAttributes;
        i: HTMLAttributes;
        iframe: IframeHTMLAttributes;
        img: ImgHTMLAttributes;
        input: InputHTMLAttributes;
        ins: InsHTMLAttributes;
        kbd: HTMLAttributes;
        keygen: KeygenHTMLAttributes;
        label: LabelHTMLAttributes;
        legend: HTMLAttributes;
        li: LiHTMLAttributes;
        link: LinkHTMLAttributes;
        main: HTMLAttributes;
        map: MapHTMLAttributes;
        mark: HTMLAttributes;
        menu: MenuHTMLAttributes;
        menuitem: HTMLAttributes;
        meta: MetaHTMLAttributes;
        meter: MeterHTMLAttributes;
        nav: HTMLAttributes;
        noscript: HTMLAttributes;
        object: ObjectHTMLAttributes;
        ol: OlHTMLAttributes;
        optgroup: OptgroupHTMLAttributes;
        option: OptionHTMLAttributes;
        output: OutputHTMLAttributes;
        p: HTMLAttributes;
        param: ParamHTMLAttributes;
        picture: HTMLAttributes;
        pre: HTMLAttributes;
        progress: ProgressHTMLAttributes;
        q: QuoteHTMLAttributes;
        rp: HTMLAttributes;
        rt: HTMLAttributes;
        ruby: HTMLAttributes;
        s: HTMLAttributes;
        samp: HTMLAttributes;
        search: HTMLAttributes;
        slot: SlotHTMLAttributes;
        script: ScriptHTMLAttributes;
        section: HTMLAttributes;
        select: SelectHTMLAttributes;
        small: HTMLAttributes;
        source: SourceHTMLAttributes;
        span: HTMLAttributes;
        strong: HTMLAttributes;
        style: StyleHTMLAttributes;
        sub: HTMLAttributes;
        summary: HTMLAttributes;
        sup: HTMLAttributes;
        table: TableHTMLAttributes;
        template: HTMLAttributes;
        tbody: HTMLAttributes;
        td: TdHTMLAttributes;
        textarea: TextareaHTMLAttributes;
        tfoot: HTMLAttributes;
        th: ThHTMLAttributes;
        thead: HTMLAttributes;
        time: TimeHTMLAttributes;
        title: HTMLAttributes;
        tr: HTMLAttributes;
        track: TrackHTMLAttributes;
        u: HTMLAttributes;
        ul: HTMLAttributes;
        var: HTMLAttributes;
        video: VideoHTMLAttributes;
        wbr: HTMLAttributes;
    }
    export {};
}